Add NMEA-over-UDP support
`ubxtool` can now forward NMEA sentences as UDP datagrams;
this can be used to feed them to `gpsd`, avoiding the need
for `gpsd` to have its own connection to the GPS receiver.

The `--udp-nmea` command line argument accepts an IPv4/6
address-and-port as the destination to receive the datagrams.

The behavior of the `--keep-nmea` command line argument has
also been changed; it now ensures that NMEA is enabled in
the receiver, instead of assuming it is enabled by default.